requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
# pydantic==2.5.0
# pydantic-settings==2.1.0

# LLM & AI (local — no cloud APIs)
ollama
chromadb==0.4.18          # legacy fallback; primary store is Qdrant

# Vector Store (primary)
qdrant-client>=1.7.0

# BM25 sparse index (Rust-backed, 10x faster than Whoosh)
# Install: pip install tantivy
tantivy>=0.20.0

# Reranking & Embeddings (local)
sentence-transformers>=2.3.0
# torch            # CPU-only is fine; swap "torch" for "torch --index-url https://download.pytorch.org/whl/cu130" depedning on CUDA version if GPU is available

# Web Crawling & HTTP
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
# lxml==4.9.3
trafilatura>=1.6.0          # Clean content extraction (better than BS4 alone)
# scrapy>=2.11.0            # Optional: swap async crawler for Scrapy in high-volume setups

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Image Analysis (OCR — also requires system Tesseract: https://github.com/UB-Mannheim/tesseract/wiki)
Pillow>=10.0.0
pytesseract>=0.3.10

# NLP
spacy>=3.7.0               # After install: python -m spacy download en_core_web_sm
langdetect>=1.0.9

# Caching (3-layer: queries / embeddings / LLM responses)
diskcache>=5.6.0

# Data Processing
pandas
numpy

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3

# Search
duckduckgo-search==3.9.6
